// src/apps/security/module/store/types.d.ts
import { IMainLoading } from "types";

export interface ILoading extends IMainLoading {}

export interface IAddOrEditModuleRequest {
  nameAz: string;
  nameEn: string;
  nameRu: string;
  shortNameAz: string;
  shortNameEn: string;
  shortNameRu: string;
  icon: string;
  url: string;
  code: string;
  applicationId: string;
  parentId: string;
  id?: string;
}